Workaround the inaccessbility of WebSessionManager.Val
The inability to mock the guava cache and also
the WebSessionManager.Val force the need to "hack" the access
to a "fake" Val by hijacking the constructor with a static
method inside the same package com.google.gerrit.httpd.

The hack allows to resume the normal execution of
the mocked tests since the introduction of the stricter
errorprone rules that denied the Guava cache mocking.

Change-Id: I1efcd1b145cae7f7921019fc21c8f34e72d5288e