Do not use persistent web_sessions
The websession-broker plugin is able to consume web_sessions from a
broker and store them to a cache.

The cached web_sessions however do not require to be persisted on disk
since since the can be populated at start time by just reading entries
from the broker again.

Stop using H2 persistent cache to store web_sessions produced and
consumed from the websession-broker and rely instead of an in-memory
only cache.

Bug: Issue 14128
Change-Id: I79ec75ca941b1b49416a6ba9f38192aeebea8105