Add config parameter to disable web sessions replay after restart
Replaying all web sessions after startup make sense when
`web_sessions` cache is an in-memory cache. For persisted caches
for example when cache.*.diskLimit is greater then zero or cache
backed implementation is persisted(for example cache-chroniclemap)
this can be set to false because we need to consume just the new
web sessions.

Add `replayAllSessions` configuration parameter to allow skipping
web sessions replay

Change-Id: I630fbdf05bac24512e7283ef0dc8531bdc8cbd88